requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
apache-airflow==2.9.2
apache-airflow-providers-common-io==1.3.2
apache-airflow-providers-common-sql==1.2.0
apache-airflow-providers-fab==1.1.1
apache-airflow-providers-ftp==3.1.0
apache-airflow-providers-http==4.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.0.0
apispec==6.6.0
argcomplete==1.12.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
attrs==23.1.0
Babel==2.11.0
backoff==2.2.1
bcrypt==3.2.0
blinker==1.6.2
Brotli==1.0.9
cachelib==0.9.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
colour==0.1.5
ConfigUpdater==3.1.1
connexion==2.14.0
cron-descriptor==1.2.24
croniter==2.0.5
cryptography==42.0.5
Deprecated==1.2.13
dill==0.3.8
dnspython==2.4.2
docutils==0.18.1
duckdb==1.0.0
email_validator==2.2.0
fastapi==0.103.0
Flask==2.2.5
Flask-AppBuilder==4.4.1
Flask-Babel==2.0.0
Flask-Caching==2.0.1
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
fsspec==2024.3.1
furl==2.1.2
google-re2==1.1.20240601
googleapis-common-protos==1.56.4
greenlet==3.0.1
grpcio==1.62.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.7
importlib-metadata==7.0.1
importlib-resources==6.1.1
infinity==1.5
inflection==0.5.1
iniconfig==1.1.1
intervals==0.9.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kafka-python==2.0.2
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.0
lockfile==0.12.2
Mako==1.2.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.0
mdurl==0.1.0
methodtools==0.4.7
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
ordered-set==4.1.0
orderedmultidict==1.0.1
packaging==23.2
passlib==1.7.4
pathspec==0.10.3
pendulum==2.1.2
phonenumbers==8.12.27
pip==24.0
pluggy==1.0.0
prison==0.2.1
prometheus_client==0.20.0
protobuf==4.25.3
psutil==5.9.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.15.1
PyJWT==2.8.0
PySocks==1.7.1
pytest==7.4.4
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-nvd3==0.15.0
python-slugify==5.0.2
pytz==2024.1
pytzdata==2020.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.32.2
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.3.5
rich-argparse==1.5.2
rpds-py==0.10.6
setproctitle==1.2.2
setuptools==69.5.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.2
sqlparse==0.4.4
starlette==0.27.0
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.1.0
text-unidecode==1.3
typing_extensions==4.11.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
Unidecode==1.2.0
universal_pathlib==0.2.2
urllib3==2.2.1
uvicorn==0.20.0
Werkzeug==2.3.8
wheel==0.43.0
wirerope==0.4.7
wrapt==1.14.1
WTForms==2.3.3
zipp==3.17.0
